src/test/java/com/au/api/location/controller/LocationControllerTest.java
package com.au.api.location.controller;

import org.junit.Before;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.mockito.junit.MockitoJUnitRunner;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
import org.springframework.test.web.servlet.setup.MockMvcBuilders;

import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;

@RunWith(MockitoJUnitRunner.class)
public class LocationControllerTest {

    private MockMvc mockMvc;

    @Before
    public void setUp() {
        mockMvc = MockMvcBuilders.standaloneSetup(new LocationController()).build();
    }

    @Test
    public void shouldReturnLocations() throws Exception {
        // Given
        
        // When
        mockMvc.perform(get("/venues/search?query="))
                .andDo(print())
                // Then
                .andExpect(status().isOk());
//                .andExpect(content().string(containsString("Hello World")));

    }
}